# perf(optimization): improve req/sec 575%